Added player health bar and direct sunlight detection
The system for the direct sunlight isn't perfect but it is mostly functional. I will need to make some further changes soon. The DirectSunlightManager will automatically generate the ShapeCast2D nodes for every shadow in the scene. I also created the PlayerStatManager which will contain all of the player's stat information.