Added support for the Parasol
Created an AnimatedSprite2D node for the parasol and setup all relevant Inputs and functionality. The current sprite sheet is a placeholder and should be replaced in the near future. As of now when the Parasol is fully opened and the player is in direct sunlight it will shield the player from the sunlight and allow the passive health regeneration to continue.