Finished Player Animation Tree
All player animation states behave as expected now (I think? lmao). I had to take the equivalent of a crash course in the AnimationTree and AnimationPlayer nodes. I wish I could delete the Start and End nodes in the AnimationTree graph because they are not useful in our case and they are just kind of there. =(

Basic Overview:
- Player can only open/close the Parasol from the Idle animation
- Player can only Jump from the Idle or Running animations
- Player can only Attack when the Parasol is closed
- Player Falls slower while the Parasol is open
- Player can Attack in alternate directions by holding the attack_modifier_up or attack_modifier_down keys (W and S by default)